% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DRM}
\alias{DRM}
\title{Calculates the realized dominance relationship matrix}
\usage{
DRM(X)
}
\arguments{
\item{X}{is a (numeric, \eqn{n}{m}) matrix of genotypes with a number of rows identical to the number of SNPs (m) and a number of columns identical to the number of genotyped individuals (n). Values in the matrix are 0 and 2 for homozygous and 1 for heterozygous.}
}
\description{
Calculates realized dominance relationship matrix.
}
\examples{
## Not to run ##

## DRM(X)

## End(Not run)

}
\references{
VanRaden, Paul M. Efficient methods to compute genomic predictions. Journal of dairy science 91.11 (2008): 4414-4423.

Su G, Christensen OF, Ostersen T, Henryon M, Lund MS. 2012. Estimating Additive and Non-Additive Genetic Variances and Predicting Genetic Merits Using Genome-Wide Dense Single Nucleotide Polymorphism Markers. PLoS ONE 7(9): e45293. doi:10.1371/journal.pone.0045293
}
