% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRM.R
\name{NRM}
\alias{NRM}
\title{Calculates the Pedigree-based Additive Relationship Matrix}
\usage{
NRM(Pedig, keep.only = NULL, keep = keep.only, AFounder = NULL)
}
\arguments{
\item{Pedig}{Data frame containing the Pedigree. The data frame has columns (1) Individual, (2) Sire, (3) Dam. Missing parents are coded as NA. Both parents must either be missing or present.}

\item{keep.only}{If \code{keep.only} is provided then kinships are computed only for these animals.}

\item{keep}{If \code{keep} is provided then kinships are computed only for these animals and their ancestors.}

\item{AFounder}{Additive relationship matrix of the founders. The row names are the ids of the founders. By default, founders are assumed to be unrelated. Founders not included in this matrix are also assumed to be unrelated.}
}
\value{
Additive relationship matrix.
}
\description{
Calculates the the Pedigree-based Additive Relationship Matrix. This is twice the pedigree based kinship matrix.
}
\examples{
## Not to run ##

## NRM(Pedigree, keep.only=NULL, keep=keep.only, AFounder=NULL)

## End(Not run)

}
